move description to zoo.hpp header, implement new fromYamlFile method for NNModelDescription
