+++ README.md
@@ -192,7 +192,8 @@ The following environment variables can be set to alter default behavior of the 
 | DEPTHAI_HUB_API_KEY | API key for the Luxonis Hub |
 | DEPTHAI_ZOO_INTERNET_CHECK | (Default) 1 - perform internet check, 0 - skip internet check |
 | DEPTHAI_ZOO_INTERNET_CHECK_TIMEOUT | (Default) 1000 - timeout in milliseconds for the internet check |
-| DEPTHAI_ZOO_CACHE_FOLDER | (Default) .depthai_cached_models - Folder where model zoo models are stored |
+| DEPTHAI_ZOO_CACHE_FOLDER | (Default) .depthai_cached_models - Folder where cached zoo models are stored |
+| DEPTHAI_ZOO_MODELS_PATH | (Default) depthai_models - Folder where zoo model description files are stored |
 | DEPTHAI_RECORD | Enables holistic record to the specified directory. |
 | DEPTHAI_REPLAY | Replays holistic replay from the specified file or directory. |
