+++ include/depthai/pipeline/node/DetectionNetwork.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <depthai/modelzoo/NNModelDescription.hpp>
+#include <depthai/modelzoo/Zoo.hpp>
 #include <depthai/pipeline/DeviceNodeGroup.hpp>
 #include <depthai/pipeline/Subnode.hpp>
 #include <depthai/pipeline/node/Camera.hpp>
