+++ tests/src/onhost_tests/model_slug_test.cpp
@@ -1,6 +1,6 @@
 #define CATCH_CONFIG_MAIN
 #include <catch2/catch_all.hpp>
-#include <depthai/modelzoo/NNModelDescription.hpp>
+#include <depthai/modelzoo/Zoo.hpp>
 #include <sstream>
 #include <string>
